requirements.txt
gradio
transformers
accelerate
torch
chromadb
langchain
sentence-transformers
pillow
pytesseract
pymupdf
python-docx
pandas
langchain-community
duckdb